Organize identifier editor code into its own package
